Update requested changes

+++ core/constants.go
@@ -10,7 +10,7 @@ import (
 var EpochLength uint64 = 1200
 var NumberOfStates uint64 = 5
 var ChainId = big.NewInt(0x785B4B9847B9)
-var StateLength = uint64(EpochLength / NumberOfStates)
+var StateLength = EpochLength / NumberOfStates
 var MaxRetries uint = 8
 var NilHash = common.Hash{0x00}
 var BlockCompletionTimeout = 30
